api-gateway/requirements.txt
flask
flask_sqlalchemy
flask_migrate
requests
flask_cors
flask_restful
python-dotenv
prometheus_flask_exporter